refactor(targets): move concrete target modules into param_decomp/targets/
Group the vendored-target impls under a `targets/` subpackage: `llama8b.py`,
`llama_simple_mlp.py`, `target_aliases.py`, and the target-coupled
`llama8b_sharding.py` (it imports `LlamaDecomposedModel`). The `DecomposedModel`
interface (`lm.py`) and the shared `components.py` primitive stay at top level.

Updated every import across core + lab + tests (30 modules) plus the CLAUDE.md /
README path references. Pure move + import rewrite; no behavior change.
