+++ param_decomp/experiments/invariance_check.py
@@ -41,13 +41,13 @@ from param_decomp.configs import (
     SCScope,
     UniformKSubsetRoutingConfig,
 )
-from param_decomp.llama8b import (
-    llama_site_specs,
-    mlp_family_site_cs,
-)
 from param_decomp.recon import build_recon_terms
 from param_decomp.schedule import ScheduleConfig
 from param_decomp.sharding import dp_mesh, shard_batch
+from param_decomp.targets.llama8b import (
+    llama_site_specs,
+    mlp_family_site_cs,
+)
 from param_decomp.tests.test_llama8b import _tiny_cfg, _tiny_decomposed_lm
 from param_decomp.train import TrainState, make_train_step
