+++ param_decomp/targets/target_aliases.py
@@ -5,8 +5,8 @@ The toy targets (TMS, ResidMLP) live in the lab and are NOT members of these uni
 the generic engine (`run_decomposition_training`) takes the model as `DecomposedModel`, so
 the core never names a toy."""
 
-from param_decomp.llama8b import LlamaDecomposedModel, Prefix
-from param_decomp.llama_simple_mlp import SimpleMLPDecomposedModel, SimpleMLPPrefix
+from param_decomp.targets.llama8b import LlamaDecomposedModel, Prefix
+from param_decomp.targets.llama_simple_mlp import SimpleMLPDecomposedModel, SimpleMLPPrefix
 
 AnyDecomposedModel = LlamaDecomposedModel | SimpleMLPDecomposedModel
 AnyPrefix = Prefix | SimpleMLPPrefix
