+++ param_decomp/tests/stacked_parity/gen_stacked_fixtures.py
@@ -42,15 +42,15 @@ from param_decomp.configs import (  # noqa: E402
     PersistentPGDReconLossConfig,
     SCScope,
 )
-from param_decomp.llama8b import (  # noqa: E402
+from param_decomp.recon import subset_chunk_plan  # noqa: E402
+from param_decomp.schedule import ScheduleConfig  # noqa: E402
+from param_decomp.targets.llama8b import (  # noqa: E402
     KINDS,
     LayerRange,
     init_decomp_vu,
     llama_decomposed_lm,
     site_name,
 )
-from param_decomp.recon import subset_chunk_plan  # noqa: E402
-from param_decomp.schedule import ScheduleConfig  # noqa: E402
 from param_decomp.tests.test_llama8b import _tiny_cfg, _tiny_target  # noqa: E402
 from param_decomp.train import TrainState, make_train_step  # noqa: E402
 from vendored_jax.llama import LlamaConfig  # noqa: E402
