+++ param_decomp/tests/test_eval.py
@@ -20,8 +20,8 @@ from param_decomp.ci_fn import (
 )
 from param_decomp.config import EvalPGDConfig
 from param_decomp.eval import make_eval_step, next_token_cross_entropy
-from param_decomp.llama8b import llama_site_specs, mlp_family_site_cs
 from param_decomp.lm import DecomposedModel, SiteSpec
+from param_decomp.targets.llama8b import llama_site_specs, mlp_family_site_cs
 from param_decomp.tests.test_llama8b import (
     _tiny_cfg,
     _tiny_decomposed_lm,
