+++ param_decomp/tests/test_fresh_pgd_cscope_dp_invariance.py
@@ -29,12 +29,12 @@ from jax import random
 
 from param_decomp.adversary import init_fresh_pgd_sources, source_masks
 from param_decomp.components import init_decomp_vu
-from param_decomp.llama8b import (
+from param_decomp.losses import kl_per_position
+from param_decomp.sharding import dp_mesh, shard_batch
+from param_decomp.targets.llama8b import (
     llama_site_specs,
     mlp_family_site_cs,
 )
-from param_decomp.losses import kl_per_position
-from param_decomp.sharding import dp_mesh, shard_batch
 from param_decomp.tests.test_llama8b import _tiny_cfg, _tiny_decomposed_lm
