+++ param_decomp/tests/test_llama_simple_mlp.py
@@ -30,8 +30,11 @@ from param_decomp.configs import (
     SCScope,
     UniformKSubsetRoutingConfig,
 )
-from param_decomp.llama8b import FrozenAttn
-from param_decomp.llama_simple_mlp import (
+from param_decomp.lm import DecomposedModel, SiteC, SiteSpec
+from param_decomp.recon import build_recon_terms
+from param_decomp.schedule import ScheduleConfig
+from param_decomp.targets.llama8b import FrozenAttn
+from param_decomp.targets.llama_simple_mlp import (
     LlamaSimpleMLPConfig,
     SimpleMLPDecomposedModel,
     SimpleMLPPrefix,
@@ -46,9 +49,6 @@ from param_decomp.llama_simple_mlp import (
     site_name,
     site_specs,
 )
-from param_decomp.lm import DecomposedModel, SiteC, SiteSpec
-from param_decomp.recon import build_recon_terms
-from param_decomp.schedule import ScheduleConfig
 from param_decomp.train import TrainState, make_faith_warmup_step, make_train_step
