+++ param_decomp_lab/experiments/lm/pretrain/CLAUDE.md
@@ -39,7 +39,7 @@ A freshly-pretrained target is decomposable with NO conversion. `pretrain.train`
 `PARAM_DECOMP_OUT_DIR/pretrain_cache/<project>-<run_id>/model_step_<N>.safetensors` keyed
 `h.{i}.attn.{q,k,v,o}_proj.weight`, `h.{i}.mlp.{c_fc,down_proj}.weight`,
 `h.{i}.rms_{1,2}.weight`, `wte.weight`, `ln_f.weight` (NO `lm_head.weight` — tied), every
-weight `(d_out, d_in)` — exactly what `param_decomp.llama_simple_mlp` reads. A
+weight `(d_out, d_in)` — exactly what `param_decomp.targets.llama_simple_mlp` reads. A
 decomposition config points at it via `target.spec`:
 
 ```yaml
